refactor(import): gather the module settings into their own cached unit
