+++ internal/plugins/import/utils/settings_test.go
@@ -63,3 +63,50 @@ func TestModuleSettingsIsExternalPath(t *testing.T) {
 		})
 	}
 }
+
+func TestModuleSettingsIsIgnoredPath(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name     string
+		settings map[string]interface{}
+		fileName string
+		want     bool
+	}{
+		{
+			name:     "array of interface strings matches as regexp",
+			settings: map[string]interface{}{"import/ignore": []interface{}{"ignored-missing-default"}},
+			fileName: "/repo/ignored-missing-default.ts",
+			want:     true,
+		},
+		{
+			name:     "array of strings matches as regexp",
+			settings: map[string]interface{}{"import/ignore": []string{`\.css$`}},
+			fileName: "/repo/styles.css",
+			want:     true,
+		},
+		{
+			name:     "non-string entries and invalid regexps are ignored",
+			settings: map[string]interface{}{"import/ignore": []interface{}{123, "["}},
+			fileName: "/repo/ignored-missing-default.ts",
+			want:     false,
+		},
+		{
+			name:     "missing setting does not ignore",
+			settings: map[string]interface{}{},
+			fileName: "/repo/ignored-missing-default.ts",
+			want:     false,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+
+			got := import_utils.CompileModuleSettings(tc.settings).IsIgnoredPath(tc.fileName)
+			if got != tc.want {
+				t.Fatalf("IsIgnoredPath() = %v, want %v", got, tc.want)
+			}
+		})
+	}
+}
